backend/backend/requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=2.9
django-cors-headers>=4.0
python-dotenv>=1.0
google-api-python-client>=2.70
google-auth>=2.15
google-auth-oauthlib>=0.8
google-auth-httplib2>=0.1